clearing all users in clearing_db_test is now conditional

+++ src/clearing_db_test.py
@@ -14,13 +14,14 @@ SOLR_SERVER = "http://xantoz.failar.nu:8080/solr/"
 
 def clear_database_and_add_users(usernames=["SSDummy_Janet", "ssdummy_henry", "ssdummy_hoot", "ssdummy_faye"
                                             ,"ssdummy_burt", "ssdummy_gustavo", "ssdummy_amanda"
-                                            ,"ssdummy_duke", "ssdummy_ian", "ssdummy_ellen" , "ssdummy_chrissy"]):
+                                            ,"ssdummy_duke", "ssdummy_ian", "ssdummy_ellen" , "ssdummy_chrissy"], clear=False):
     '''Used by analyse team to test their accuracy!  Removes all users
     from the database and then adds the ones in the usernames list.
     TODO: test it'''
     
     sh = StorageHandler(SOLR_SERVER)
-    sh.delete_all()
+    if clear:
+        sh.delete_all()
     users_left_to_add = len(usernames)
     print "CLEARING DATABASE!"
     for username in usernames:
@@ -34,4 +35,4 @@ def clear_database_and_add_users(usernames=["SSDummy_Janet", "ssdummy_henry", "s
     
 # testing
 if __name__ == '__main__':
-    clear_database_and_add_users()
+    clear_database_and_add_users(clear=True)
